+++ actify/src/handles/mod.rs
@@ -1,7 +1,10 @@
+mod builder;
 mod handle;
 mod read_handle;
 
 #[cfg(test)]
-pub(crate) use handle::CHANNEL_SIZE;
+pub(crate) use builder::DEFAULT_BROADCAST_CAPACITY;
+
+pub use builder::HandleBuilder;
 pub use handle::{BroadcastAs, Handle};
 pub use read_handle::ReadHandle;